Added floats to title lengths

+++ backend/database/query_gettitlelength.go
@@ -30,8 +30,8 @@ func (db *DB) GetTitleLength(w http.ResponseWriter, r *http.Request) {
 		// Output structure
 		var output struct {
 			Years   []int `json:"years"`
-			Title_1 []int `json:"title_1"`
-			Title_2 []int `json:"title_2"`
+			Title_1 []float64 `json:"title_1"`
+			Title_2 []float64 `json:"title_2"`
 		}
 
 		// Grab input values from url
@@ -104,8 +104,8 @@ func (db *DB) GetTitleLength(w http.ResponseWriter, r *http.Request) {
 		defer rows.Close()
 		var (
 			year    int
-			title_1 int
-			title_2 int
+			title_1 float64 
+			title_2 float64 
 		)
 		for rows.Next() {
 			// Each row's values are put in temporary variables
